# Ride-Now/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RideNow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  driver: "driver",
  admin: "admin",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
  /** @nullable */
  phone?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  driver: "driver",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
}

export type RideRideType = (typeof RideRideType)[keyof typeof RideRideType];

export const RideRideType = {
  cab: "cab",
  bike: "bike",
} as const;

export type RideStatus = (typeof RideStatus)[keyof typeof RideStatus];

export const RideStatus = {
  requested: "requested",
  accepted: "accepted",
  ongoing: "ongoing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Ride {
  id: number;
  customerId: number;
  /** @nullable */
  driverId?: number | null;
  pickupAddress: string;
  dropoffAddress: string;
  rideType: RideRideType;
  status: RideStatus;
  /** @nullable */
  fare?: number | null;
  /** @nullable */
  distance?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  driverName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateRideBodyRideType =
  (typeof CreateRideBodyRideType)[keyof typeof CreateRideBodyRideType];

export const CreateRideBodyRideType = {
  cab: "cab",
  bike: "bike",
} as const;

export interface CreateRideBody {
  pickupAddress: string;
  dropoffAddress: string;
  rideType: CreateRideBodyRideType;
  distance: number;
}

export type UpdateRideStatusBodyStatus =
  (typeof UpdateRideStatusBodyStatus)[keyof typeof UpdateRideStatusBodyStatus];

export const UpdateRideStatusBodyStatus = {
  accepted: "accepted",
  ongoing: "ongoing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateRideStatusBody {
  status: UpdateRideStatusBodyStatus;
}

export type BookingBookingType =
  (typeof BookingBookingType)[keyof typeof BookingBookingType];

export const BookingBookingType = {
  rental: "rental",
  self_drive: "self_drive",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  customerId: number;
  /** @nullable */
  vehicleId?: number | null;
  bookingType: BookingBookingType;
  startDate: string;
  endDate: string;
  /** @nullable */
  totalAmount?: number | null;
  status: BookingStatus;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  vehicleInfo?: string | null;
  createdAt: string;
}

export type CreateBookingBodyBookingType =
  (typeof CreateBookingBodyBookingType)[keyof typeof CreateBookingBodyBookingType];

export const CreateBookingBodyBookingType = {
  rental: "rental",
  self_drive: "self_drive",
} as const;

export interface CreateBookingBody {
  /** @nullable */
  vehicleId?: number | null;
  bookingType: CreateBookingBodyBookingType;
  startDate: string;
  endDate: string;
}

export interface Driver {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  isOnline: boolean;
  totalEarnings: number;
  totalRides: number;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  vehicleInfo?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  createdAt: string;
}

export interface UpdateDriverBody {
  isOnline?: boolean;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  vehicleInfo?: string | null;
}

export interface EarningsSummary {
  totalEarnings: number;
  todayEarnings: number;
  weekEarnings: number;
  totalRides: number;
  completedRides: number;
  /** @nullable */
  averageRating?: number | null;
}

export type VehicleVehicleType =
  (typeof VehicleVehicleType)[keyof typeof VehicleVehicleType];

export const VehicleVehicleType = {
  cab: "cab",
  bike: "bike",
  suv: "suv",
  van: "van",
} as const;

export interface Vehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  vehicleType: VehicleVehicleType;
  isAvailable: boolean;
  /** @nullable */
  ratePerHour?: number | null;
  /** @nullable */
  ratePerKm?: number | null;
  /** @nullable */
  driverId?: number | null;
  createdAt: string;
}

export type CreateVehicleBodyVehicleType =
  (typeof CreateVehicleBodyVehicleType)[keyof typeof CreateVehicleBodyVehicleType];

export const CreateVehicleBodyVehicleType = {
  cab: "cab",
  bike: "bike",
  suv: "suv",
  van: "van",
} as const;

export interface CreateVehicleBody {
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  vehicleType: CreateVehicleBodyVehicleType;
  /** @nullable */
  ratePerHour?: number | null;
  /** @nullable */
  ratePerKm?: number | null;
  /** @nullable */
  driverId?: number | null;
}

export interface UpdateVehicleBody {
  isAvailable?: boolean;
  /** @nullable */
  ratePerHour?: number | null;
  /** @nullable */
  ratePerKm?: number | null;
  /** @nullable */
  driverId?: number | null;
}

export interface AdminDashboard {
  totalUsers: number;
  totalDrivers: number;
  activeDrivers: number;
  totalRides: number;
  completedRides: number;
  cancelledRides: number;
  pendingRides: number;
  totalRevenue: number;
  todayRevenue: number;
  totalVehicles: number;
}

export interface DailyRideStats {
  date: string;
  count: number;
  revenue: number;
}

export interface RideTypeStats {
  rideType: string;
  count: number;
}

export interface RideStatusStats {
  status: string;
  count: number;
}

export interface Analytics {
  dailyRides: DailyRideStats[];
  ridesByType: RideTypeStats[];
  ridesByStatus: RideStatusStats[];
}

export type GetRidesParams = {
  status?: string;
  /**
   * @nullable
   */
  customerId?: number | null;
  /**
   * @nullable
   */
  driverId?: number | null;
};