# Ride-Now/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RideNow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get bookings (rentals/self-drive)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (rental/self-drive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /drivers/{id}/accept-ride/{rideId}:
    patch:
      operationId: acceptRide
      tags: [drivers]
      summary: Driver accepts a ride
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: rideId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
...
  /admin/drivers/{id}:
    patch:
      operationId: updateDriver
      tags: [admin]
      summary: Update driver (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        phone:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Ride:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        rideType:
          type: string
          enum: [cab, bike]
        status:
          type: string
          enum: [requested, accepted, ongoing, completed, cancelled]
        fare:
          type: ["number", "null"]
        distance:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - pickupAddress
        - dropoffAddress
        - rideType
        - status
        - createdAt
        - updatedAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        vehicleId:
          type: ["integer", "null"]
        bookingType:
          type: string
          enum: [rental, self_drive]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, confirmed, active, completed, cancelled]
        customerName:
          type: ["string", "null"]
        vehicleInfo:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - bookingType
        - startDate
        - endDate
        - status
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        vehicleId:
          type: ["integer", "null"]
        bookingType:
          type: string
          enum: [rental, self_drive]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required:
        - bookingType
        - startDate
        - endDate
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        isOnline:
          type: boolean
        totalEarnings:
          type: number
        totalRides:
          type: integer
        rating:
          type: ["number", "null"]
        vehicleInfo:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - email
        - isOnline
        - totalEarnings
        - totalRides
        - createdAt
    UpdateDriverBody:
      type: object
      properties:
        isOnline:
          type: boolean
        licenseNumber:
          type: ["string", "null"]
        vehicleInfo:
          type: ["string", "null"]
...