# Ride-Now/lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RideNow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyBookingType } from "./createBookingBodyBookingType";

export interface CreateBookingBody {
  /** @nullable */
  vehicleId?: number | null;
  bookingType: CreateBookingBodyBookingType;
  startDate: Date;
  endDate: Date;
}