# Ride-Now/lib/api-zod/src/generated/types/driver.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RideNow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Driver {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  isOnline: boolean;
  totalEarnings: number;
  totalRides: number;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  vehicleInfo?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  createdAt: Date;
}