# Ride-Now/lib/api-zod/src/generated/types/updateDriverBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RideNow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDriverBody {
  isOnline?: boolean;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  vehicleInfo?: string | null;
}